+++ EngineCore/Include/SimpleEngine/ECS/Components/ChildrenComponent.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "SimpleEngine/Core/Container/Array.h"
+#include "SimpleEngine/Core/Reflection/Annotations.h"
 #include "SimpleEngine/ECS/Entity.h"
-#include "SimpleEngine/Meta/Annotations.h"
 
 
 namespace se
