+++ EngineCore/Include/SimpleEngine/ECS/Components/ParentComponent.h
@@ -1,6 +1,6 @@
 #pragma once
+#include "SimpleEngine/Core/Reflection/Annotations.h"
 #include "SimpleEngine/ECS/World.h"
-#include "SimpleEngine/Meta/Annotations.h"
 
 
 namespace se
