+++ EngineCore/Include/SimpleEngine/ECS/Components/TransformComponent.h
@@ -1,7 +1,6 @@
 #pragma once
-#include "SimpleEngine/Core/HAL/PlatformTypes.h"
 #include "SimpleEngine/Core/Math/Math.h"
-#include "SimpleEngine/Meta/Annotations.h"
+#include "SimpleEngine/Core/Reflection/Annotations.h"
 
 
 namespace se
